+++ src/components/index.js
@@ -1,0 +1,25 @@
+import Button from './Button/Button';
+import Card from './Card/Card';
+import CardNumberInput from './CardNumberInput/CardNumberInput';
+import CardSelector from './CardSelector/CardSelector';
+import ErrorMessageBox from './ErrorMessageBox/ErrorMessageBox';
+import Header from './Header/Header';
+import Input from './Input/Input';
+import InputBox from './InputBox/InputBox';
+import Modal from './Modal/Modal';
+import PinNumberInput from './PinNumberInput/PinNumberInput';
+import ToolTip from './ToolTip/ToolTip';
+
+export {
+  Button,
+  Card,
+  CardNumberInput,
+  CardSelector,
+  ErrorMessageBox,
+  Header,
+  Input,
+  InputBox,
+  Modal,
+  PinNumberInput,
+  ToolTip,
+};
