+++ src/constants/index.js
@@ -1,0 +1,7 @@
+import CARD from './card';
+import LOCAL_STORAGE_KEY from './localStorageKey';
+import MESSAGE from './message';
+import REGEX from './regex';
+import ROUTE from './route';
+
+export { CARD, LOCAL_STORAGE_KEY, MESSAGE, REGEX, ROUTE };
