+++ src/hooks/index.js
@@ -1,0 +1,6 @@
+import useInput from './useInput';
+import useLocalStorage from './useLocalStorage';
+import useModal from './useModal';
+import useMultipleInput from './useMultipleInput';
+
+export { useInput, useLocalStorage, useModal, useMultipleInput };
